+++ couplet_composer/support/project_values.py
@@ -17,74 +17,68 @@ from .project_names import get_ode_repository_name
 
 
 @cached
-def _get_project_values_file(source_root):
+def _get_project_values_file(source_root, in_tree_build):
     """
     Gives the path to the file that contains the project values.
 
     source_root -- Path to the directory that is the root of the
     script run.
+
+    in_tree_build -- Whether the build files are created in tree.
     """
-    return os.path.join(
-        source_root,
-        get_ode_repository_name(),
-        get_project_values_file_path()
-    )
+    return os.path.join(source_root, get_project_values_file_path()) \
+        if in_tree_build \
+        else os.path.join(
+            source_root,
+            get_ode_repository_name(),
+            get_project_values_file_path()
+        )
 
 
 @cached
-def _get_project_values(source_root):
+def _get_project_values(source_root, in_tree_build):
     """
     Gives the project values.
 
     source_root -- Path to the directory that is the root of the
     script run.
+
+    in_tree_build -- Whether the build files are created in tree.
     """
     try:
-        with open(_get_project_values_file(source_root=source_root)) as f:
+        with open(_get_project_values_file(
+            source_root=source_root,
+            in_tree_build=in_tree_build
+        )) as f:
             return json.load(f)
     except Exception:
         return None
 
 
-@cached
-def get_ode_name(source_root):
-    """
-    Gives the name of Obliging Ode.
+def get_ode_name():
+    """Gives the name of Obliging Ode."""
+    return "Obliging Ode"
 
-    source_root -- Path to the directory that is the root of the
-    script run.
-    """
-    project_values = _get_project_values(source_root=source_root)
-    if project_values:
-        return project_values["ode"]["name"]
-    else:
-        return "ode-name-file-not-found"
+
+def get_anthem_name():
+    """Gives the name of Unsung Anthem."""
+    return "Unsung Anthem"
 
 
 @cached
-def get_anthem_name(source_root):
-    """
-    Gives the name of Unsung Anthem.
-
-    source_root -- Path to the directory that is the root of the
-    script run.
-    """
-    project_values = _get_project_values(source_root=source_root)
-    if project_values:
-        return project_values["anthem"]["name"]
-    else:
-        return "anthem-name-file-not-found"
-
-
-@cached
-def get_ode_version(source_root):
+def get_ode_version(source_root, in_tree_build):
     """
     Gives the default version of Obliging Ode.
 
     source_root -- Path to the directory that is the root of the
     script run.
+
+    in_tree_build -- Whether the build files are created in tree.
     """
-    project_values = _get_project_values(source_root=source_root)
+    project_values = _get_project_values(
+        source_root=source_root,
+        in_tree_build=in_tree_build
+    )
     if project_values:
         return project_values["ode"]["version"]
     else:
@@ -92,97 +86,60 @@ def get_ode_version(source_root):
 
 
 @cached
-def get_anthem_version(source_root):
+def get_anthem_version(source_root, in_tree_build):
     """
     Gives the default version of Unsung Anthem.
 
     source_root -- Path to the directory that is the root of the
     script run.
+
+    in_tree_build -- Whether the build files are created in tree.
     """
-    project_values = _get_project_values(source_root=source_root)
+    project_values = _get_project_values(
+        source_root=source_root,
+        in_tree_build=in_tree_build
+    )
     if project_values:
         return project_values["anthem"]["version"]
     else:
         return "anthem-version-file-not-found"
 
 
-@cached
-def get_opengl_version(source_root):
-    """
-    Gives the default version of OpenGL.
-
-    source_root -- Path to the directory that is the root of the
-    script run.
-    """
-    project_values = _get_project_values(source_root=source_root)
-    if project_values:
-        return project_values["opengl"]["version"]
-    else:
-        return "3.2"
-
-
-@cached
-def get_default_ode_window_name(source_root):
+def get_default_ode_window_name():
     """
     Gives the default name of the window of Obliging Ode.
-
-    source_root -- Path to the directory that is the root of the
-    script run.
     """
-    return "ode-window"
+    return get_ode_name()
 
 
-@cached
-def get_default_anthem_window_name(source_root):
+def get_default_anthem_window_name():
     """
     Gives the default name of the window of Unsung Anthem.
-
-    source_root -- Path to the directory that is the root of the
-    script run.
     """
-    return get_anthem_name(source_root=source_root)
+    return get_anthem_name()
 
 
-@cached
-def get_default_ode_logger_name(source_root):
-    """
-    Gives the default name of the logger of Obliging Ode.
-
-    source_root -- Path to the directory that is the root of the
-    script run.
-    """
-    return "ode"
-
-
-@cached
-def get_default_anthem_logger_name(source_root):
-    """
-    Gives the default name of the logger of Unsung Anthem.
-
-    source_root -- Path to the directory that is the root of the
-    script run.
-    """
-    return get_anthem_name(source_root=source_root)
-
-
-@cached
-def get_ode_binaries_base_name(source_root):
+def get_ode_binaries_base_name():
     """
     Gives the default base name for the binaries of Obliging Ode.
-
-    source_root -- Path to the directory that is the root of the
-    script run.
     """
     return "ode"
 
 
-@cached
-def get_anthem_binaries_base_name(source_root):
+def get_anthem_binaries_base_name():
     """
     Gives the default base name for the binaries of Unsung
     Anthem.
-
-    source_root -- Path to the directory that is the root of the
-    script run.
     """
     return "anthem"
+
+
+@cached
+def get_scripts_base_directory_name(coverage):
+    """
+    Gives the name of the base directory where the Lua scripts
+    are copied to.
+
+    coverage -- Whether or not code coverage is enabled.
+    """
+    return "scripts_lib" if coverage else "lib"